moved convert to dna.go
